Add Flask app and refactor analysis modules
