+++ templates/index.html
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title>Mutual Fund Dashboard</title>
+    <script src="https://cdn.plot.ly/plotly-latest.min.js"></script>
+</head>
+<body>
+    <h1>Mutual Fund Dashboard</h1>
+    <form action="/nav" method="get">
+        <label for="code">Scheme Code:</label>
+        <input type="text" id="code" name="code" />
+        <button type="submit">Show NAV</button>
+    </form>
+</body>
+</html>
