+++ templates/nav.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title>NAV Analysis</title>
+    <script src="https://cdn.plot.ly/plotly-latest.min.js"></script>
+</head>
+<body>
+    <h1>NAV Analysis</h1>
+    {{ graph_html|safe }}
+</body>
+</html>
